requirements.txt
telethon
Pillow
cryptg
python-dotenv
requests
google-generativeai
